refactor(string): UUID生成機能の簡素化
