+++ src/modules/string/index.ts
